refactor(warp_runner): split command into different files
